mailchimpform component, hide overflow in mobile, and emoji
